feat: moved auth to packages
