+++ apps/todo-graphql/package.json
@@ -31,22 +31,16 @@
   },
   "dependencies": {
     "@apollo/server": "^4.7.0",
-    "@clerk/clerk-sdk-node": "^4.8.6",
     "@nestjs/apollo": "^11.0.5",
     "@nestjs/common": "^9.0.0",
     "@nestjs/config": "^2.3.1",
     "@nestjs/core": "^9.0.0",
     "@nestjs/graphql": "^11.0.5",
-    "@nestjs/passport": "^9.0.3",
-    "@nestjs/platform-express": "^9.0.0",
     "graphql": "^16.6.0",
     "graphql-tools": "^8.3.20",
-    "jsonwebtoken": "^9.0.0",
-    "jwks-rsa": "^3.0.1",
-    "passport": "^0.6.0",
-    "passport-jwt": "^4.0.1",
     "reflect-metadata": "^0.1.13",
-    "rxjs": "^7.2.0"
+    "rxjs": "^7.2.0",
+    "@package/auth": "*"
   },
   "devDependencies": {
     "@framework/jest": "*",
