+++ apps/todo-graphql/src/app.module.ts
@@ -4,11 +4,10 @@ import { AppService } from './app.service';
 import { TodoService } from './todo.service';
 import { GraphQLModule } from '@nestjs/graphql';
 import { TodoResolver } from './todo.resolver';
-import { JwtAuthGuard } from './auth.guard';
-import { JwtStrategy } from './jwt.strategy';
 import { ConfigModule } from '@nestjs/config';
 import { ApolloServerPluginLandingPageLocalDefault } from '@apollo/server/plugin/landingPage/default';
 import { PrismaService } from './prisma.service';
+import { JwtAuthGuard, JwtStrategy } from '@package/auth/dist';
 import {
   ApolloFederationDriver,
   ApolloFederationDriverConfig,
