--- apps/todo-graphql/src/jwt.strategy.ts
@@ -1,25 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { PassportStrategy } from '@nestjs/passport';
-import { ExtractJwt, Strategy } from 'passport-jwt';
-import * as jwksRsa from 'jwks-rsa';
-
-@Injectable()
-export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor() {
-    super({
-      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-      secretOrKeyProvider: jwksRsa.passportJwtSecret({
-        cache: true,
-        rateLimit: true,
-        jwksRequestsPerMinute: 5,
-        jwksUri: process.env.CLERK_JWSK_URL || '',
-      }),
-      issuer: process.env.CLERK_ISSUER,
-      algorithms: ['RS256'],
-    });
-  }
-
-  async validate(payload: any) {
-    return payload;
-  }
-}
