+++ packages/auth/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@package/auth",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "build": "tsc",
+    "postinstall": "npm run build"
+  },
+  "dependencies": {
+    "@nestjs/graphql": "^11.0.6",
+    "@nestjs/passport": "^9.0.3",
+    "eslint-config-custom": "*",
+    "jwks-rsa": "^3.0.1",
+    "passport-jwt": "^4.0.1",
+    "tsconfig": "*"
+  },
+  "publishConfig": {
+    "access": "restricted"
+  }
+}
