+++ packages/auth/src/current-user.decorator.ts
@@ -1,4 +1,4 @@
-import { createParamDecorator } from '@nestjs/common';
+import { createParamDecorator } from "@nestjs/common";
 
 export const CurrentUser = createParamDecorator((data, host) => {
   const [root, args, ctx, info] = host.args;
