+++ packages/auth/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./jwt-auth.guard";
+export * from "./jwt.strategy";
+export * from "./current-user.decorator";
