+++ packages/auth/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@tooling/typescript/nestjs.json",
+  "include": ["src/**/*"],
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "exclude": ["node_modules"]
+}
